:seedling: Add categories seeders

+++ backend/src/database/migrations/20200428175047-create-monetary-donations.js
@@ -7,7 +7,7 @@ module.exports = {
       autoIncrement: true,
     },
     value: {
-      type: Sequelize.DOUBLE(11, 2),
+      type: Sequelize.DOUBLE,
       allowNull: false,
     },
     created_at: Sequelize.DATE,
